refactor: move `AppTester` into `test/helper`
