+++ test/village.e2e-spec.ts
@@ -1,4 +1,4 @@
-import { AppTester } from '~/utils/helpers/app-tester';
+import { AppTester } from './helper/app-tester';
 
 describe('Village (e2e)', () => {
   const baseUrl = '/villages';
